desktop-agent/requirements.txt
websockets>=12.0
mss>=9.0.1
Pillow>=10.0.0
